Move getTextFromEnumValue to hedvig-ui utils
