+++ src/features/tools/campaign-codes/components/CreateCampaignCode.tsx
@@ -1,10 +1,10 @@
 import {
-  getTextFromEnumValue,
   InfoContainer,
   SearchableDropdown,
   Spacing,
   ThirdLevelHeadline,
 } from '@hedvig-ui'
+import { getTextFromEnumValue } from '@hedvig-ui/utils/text'
 import { VisibleNoDiscountForm } from 'features/tools/campaign-codes/forms/VisibleNoDiscountForm'
 import { CreatableIncentiveTypes } from 'features/tools/campaign-codes/utils'
 import React from 'react'
